feat: Refactor Dockerfile
Modify the Dockerfile to build an image that (a) does not contain the build
dependencies and (b) configures `nom` as the ENTRYPOINT. This permits
us to use it like this:

    docker run -it --rm nom

Or:

    docker run -it --rm nom -f http://example.com/feed.rss

Or even:

    docker run -it --rm -v ~/.config/nom:/config/nom nom

We also update the Dockerfile to follow best practices w/r/t managing Go
dependencies: by installing dependencies before the compile step, we allow
the container runtime to cache the dependencies and avoid re-installing
them on every build.

The workflow builds images for amd64, arm/v7, and arm64.