ci: Build container image on push to master
We use `docker/metadata-action` to extract tags for the image. A push to
`master` will tag the image with:

- master
- <short commit hash>

A push to a tag like `v2.16.0` will tag the image with:

- 2.16.0
- 2.16
- 2
- latest

This allows someone to pin their container to the major version
(`ghcr.io/guyfedwards/nom:2`) so that they receive updates but not breaking
changes.