fix: Fix config path handling
The config handling code was erroneously including the path separator in
arguments to `filepath.Join`, like this:

    configFile := filepath.Join(configDir, "/config.yml")

When running nom against a config file in the current directory:

    nom --config-path config.yml

This would result in `configDir` equal to `""`, and `configFile` would get
set to `/config.yml`. This commit updates the path handling code to
correctly use `filepath.Join`. It also replaces string splitting on paths
with `filepath.Split`.

We also add a `-c` short option for setting the configuration path:

    nom -c config.yml