fix: Include commit and build date in version information
When running `make`, collect the date and current commit reference and
embed them into the compiled binary. This makes it much easier to identify
exactly what state of the repository was used to generate a given binary.

With this change, the output of `nom version` looks like this when built
from an arbitrary commit:

    dev (253846e130) on 2025-10-24T12:07:18

When built from a tagged commit, it looks something like:

    v2.21.0 (53e757869a) on 2025-10-24T12:38:08